--- a/db/migrations/postgres/000044_add_batch_node.down.sql
+++ b/db/migrations/postgres/000044_add_batch_node.down.sql
@@ -1,0 +1,3 @@
+BEGIN;
+ALTER TABLE batches DROP COLUMN node_id;
+COMMIT;