--- a/db/migrations/postgres/000044_add_batch_node.up.sql
+++ b/db/migrations/postgres/000044_add_batch_node.up.sql
@@ -1,0 +1,3 @@
+BEGIN;
+ALTER TABLE batches ADD COLUMN node_id UUID;
+COMMIT;