--- a/db/migrations/sqlite/000044_add_batch_node.down.sql
+++ b/db/migrations/sqlite/000044_add_batch_node.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE batches DROP COLUMN node_id;