--- a/db/migrations/sqlite/000044_add_batch_node.up.sql
+++ b/db/migrations/sqlite/000044_add_batch_node.up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE batches ADD COLUMN node_id UUID;