--- a/internal/assets/token_transfer.go
+++ b/internal/assets/token_transfer.go
@@ -284,7 +284,7 @@ func (s *transferSender) sendInternal(ctx context.Context, method sendMethod) er
 		}
 		if s.transfer.Message != nil {
 			s.transfer.Message.State = fftypes.MessageStateStaged
-			err = s.mgr.database.UpsertMessage(ctx, &s.transfer.Message.Message, false, false)
+			err = s.mgr.database.UpsertMessage(ctx, &s.transfer.Message.Message, database.UpsertOptimizationNew)
 		}
 		return err
 	})