--- a/internal/broadcast/message.go
+++ b/internal/broadcast/message.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/hyperledger/firefly/internal/i18n"
 	"github.com/hyperledger/firefly/internal/sysmessaging"
+	"github.com/hyperledger/firefly/pkg/database"
 	"github.com/hyperledger/firefly/pkg/fftypes"
 )
 
@@ -157,7 +158,7 @@ func (s *broadcastSender) sendInternal(ctx context.Context, method sendMethod) (
 	}
 
 	// Store the message - this asynchronously triggers the next step in process
-	return s.mgr.database.UpsertMessage(ctx, &s.msg.Message, false, false)
+	return s.mgr.database.UpsertMessage(ctx, &s.msg.Message, database.UpsertOptimizationNew)
 }
 
 func (s *broadcastSender) isRootOrgBroadcast(ctx context.Context) bool {