--- a/internal/database/sqlcommon/batch_sql.go
+++ b/internal/database/sqlcommon/batch_sql.go
@@ -42,6 +42,7 @@ var (
 		"confirmed",
 		"tx_type",
 		"tx_id",
+		"node_id",
 	}
 	batchFilterFieldMap = map[string]string{
 		"type":             "btype",
@@ -49,6 +50,7 @@ var (
 		"transaction.type": "tx_type",
 		"transaction.id":   "tx_id",
 		"group":            "group_hash",
+		"node":             "node_id",
 	}
 )
 
@@ -98,6 +100,7 @@ func (s *SQLCommon) UpsertBatch(ctx context.Context, batch *fftypes.Batch, allow
 				Set("confirmed", batch.Confirmed).
 				Set("tx_type", batch.Payload.TX.Type).
 				Set("tx_id", batch.Payload.TX.ID).
+				Set("node_id", batch.Node).
 				Where(sq.Eq{"id": batch.ID}),
 			func() {
 				s.callbacks.UUIDCollectionNSEvent(database.CollectionBatches, fftypes.ChangeEventTypeUpdated, batch.Namespace, batch.ID)
@@ -124,6 +127,7 @@ func (s *SQLCommon) UpsertBatch(ctx context.Context, batch *fftypes.Batch, allow
 					batch.Confirmed,
 					batch.Payload.TX.Type,
 					batch.Payload.TX.ID,
+					batch.Node,
 				),
 			func() {
 				s.callbacks.UUIDCollectionNSEvent(database.CollectionBatches, fftypes.ChangeEventTypeCreated, batch.Namespace, batch.ID)
@@ -152,6 +156,7 @@ func (s *SQLCommon) batchResult(ctx context.Context, row *sql.Rows) (*fftypes.Ba
 		&batch.Confirmed,
 		&batch.Payload.TX.Type,
 		&batch.Payload.TX.ID,
+		&batch.Node,
 	)
 	if err != nil {
 		return nil, i18n.WrapError(ctx, err, i18n.MsgDBReadErr, "batches")