One of the earliest findings in #316 @shorsher is that the speed at which messages can be injected into FireFly using a parallel loading test runner, is significantly faster than the speed at which batches can be processed.

This is not surprising due to one process currently being parallel, and one being sequential. That's something we can look at parallelizing on the back-end later.

However, the most immediately actionable finding from log analysis was just how many DB operations we were doing.

On a simple Laptop test, it was 3-5ms per PostgreSQL request (in a single TX), so that adds up fast!

We found with the first example analyzed:
- There were 120 messages in the batch
- Overall processing of the batch was about 4 seconds
- The DB operations to load the messages (x120) + data (x120) into the database took 3/4 of those seconds
- Each DB operation was an `Upsert` meaning
   - Message ops took 4 DB calls - `SELECT` message, `UPDATE` message, `DELETE` datarefs, `INSERT` datarefs
   - Data ops took 2 DB calls - `SELECT` data, `UPDATE` data

So the first improvement in the hardening phase, is to reduce these calls, by optimizing for the case we know we're in.
These optimizations are implemented in SQL, but in a way a similar optimization should be possible in NoSQL DBs once implemented.

- Add a `node` hint to batches, to know which node sent them
- Change `message` and `data` signatures on the DB layer to take in an `UpsertOptimization ` option
  - `UpsertOptimizationSkip` - no optimization - same as today
  - `UpsertOptimizationNew` - optimize for `INSERT`
  - `UpsertOptimizationExisting` - optimize for `UPDATE`
- Removed the options for:
  - `allowExisting` - logic should always allow for this... given
  - `allowHashUpdate` - now always false - the hash of `message` and `data` cannot change on any code path
- If we are the sending node, then optimize for `UPDATE`
  - We use the `hash` as a key when updating, and check a row is updated
  - For the optimization to work **we require the DB to return `rowsUpdated`** (well adopted we believe)
  - Note for `message` the `datarefs` are fixed, so on the optimized path we don't touch them on update
- If we are not the sending node, then optimize for `INSERT`
  - Just do it, and fall back to error checking if it fails